+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tests for "Bulgarian solitaire".
 - Tests for "Longest road".
+- Tests for "Bruce lee".
 
 ## [1.14.0] - 2022-10-31
 ### Added